Add depth buffer to OpenXR with new layer system
A lot of refactoring later and I think I'm now finally happy with the flexibility and simplicity of how a frame is rendered and organized.
I added a way to have 3D and 2D layers, although no 2D layers are being captured at the moment. These will be used for the HUD and inventory/2D screens.

It's now also capturing the depth buffer of the 3D scenes in the game, which should make the interpolation at below full framerate and hopefully even the 3D effect a lot better.